move modem modem setter
